+++ apps/extension/src/libs/event.ts
@@ -42,10 +42,8 @@ export interface PupUpEvents {
     proxyChanged: ProxyConfiguration;
 
     userProperties: UserProperties;
-    userPropertiesSdk: UserProperties;
     locations: string;
     trackEvent: TrackEvent;
-    trackEventSdk: TrackEvent;
 }
 
 export interface PayloadRequest<P = any> {
