+++ apps/extension/src/libs/service/backgroundPopUpService.ts
@@ -5,10 +5,9 @@
 
 import browser from 'webextension-polyfill';
 import { backgroundEventsEmitter, NotificationData, popUpEventEmitter } from '../event';
-import { Aptabase } from '@tonkeeper/uikit/dist/hooks/analytics';
 import { UserIdentityService } from '@tonkeeper/core/dist/user-identity';
 import { ExtensionStorage } from '../storage';
-import { AptabaseSdkBackground } from '../aptabase-sdk-background';
+import { AptabaseBackground } from '../aptabase-background';
 
 let popUpPort: browser.Runtime.Port;
 const portMessagesQueue: any[] = [];
@@ -60,12 +59,12 @@ popUpEventEmitter.on('proxyChanged', message => {
 
 // End of proxy messages
 
-let aptabase: Aptabase;
+let aptabase: AptabaseBackground;
 const userIdentity = new UserIdentityService(new ExtensionStorage());
 
 popUpEventEmitter.on('userProperties', message => {
     const { aptabaseEndpoint, aptabaseKey, ...restParams } = message.params;
-    aptabase = new Aptabase({
+    aptabase = new AptabaseBackground({
         host: aptabaseEndpoint,
         key: aptabaseKey,
         appVersion: browser.runtime.getManifest().version,
@@ -77,22 +76,4 @@ popUpEventEmitter.on('userProperties', message => {
 
 popUpEventEmitter.on('trackEvent', message => {
     aptabase?.track(message.params.name, message.params.params);
-});
-
-let aptabaseSdk: AptabaseSdkBackground;
-
-popUpEventEmitter.on('userPropertiesSdk', message => {
-    const { aptabaseEndpoint, aptabaseKey, ...restParams } = message.params;
-    aptabaseSdk = new AptabaseSdkBackground({
-        host: aptabaseEndpoint,
-        key: aptabaseKey,
-        appVersion: browser.runtime.getManifest().version,
-        userIdentity
-    });
-
-    aptabaseSdk.init(restParams);
-});
-
-popUpEventEmitter.on('trackEventSdk', message => {
-    aptabaseSdk?.track(message.params.name, message.params.params);
 });
