+++ packages/uikit/src/hooks/analytics/aptabase.ts
@@ -1,17 +1,12 @@
+import { init, trackEvent } from '@aptabase/web';
 import { Network } from '@tonkeeper/core/dist/entries/network';
-import { Analytics, getUserIdentityProps } from './common';
 import { Account } from '@tonkeeper/core/dist/entries/account';
 import { UserIdentity } from '@tonkeeper/core/dist/user-identity';
 import { AnalyticsEvent } from '@tonkeeper/core/dist/analytics';
+import { Analytics, getUserIdentityProps } from './common';
 
 export class Aptabase implements Analytics {
     private user_properties: Record<string, string | number | boolean> = {};
-
-    private readonly apiUrl: string;
-
-    private readonly appKey: string;
-
-    private readonly appVersion: string;
 
     private readonly userIdentity: UserIdentity;
 
@@ -21,9 +16,10 @@ export class Aptabase implements Analytics {
         appVersion: string;
         userIdentity: UserIdentity;
     }) {
-        this.apiUrl = `${options.host}/api/v0/event`;
-        this.appKey = options.key;
-        this.appVersion = options.appVersion;
+        init(options.key, {
+            apiUrl: `${options.host}/api/v0/event`,
+            appVersion: options.appVersion
+        });
         this.userIdentity = options.userIdentity;
         this.track = this.track.bind(this);
     }
@@ -40,7 +36,6 @@ export class Aptabase implements Analytics {
         this.user_properties.walletType = params.walletType;
         this.user_properties.network = params.network === Network.TESTNET ? 'testnet' : 'mainnet';
         this.user_properties.accounts = params.accounts?.length ?? 0;
-        this.user_properties.version = this.appVersion;
         if (params.platform) {
             this.user_properties.platform = params.platform;
         }
@@ -48,102 +43,26 @@ export class Aptabase implements Analytics {
 
     track(name: string, params?: Record<string, string | number | boolean>): Promise<void>;
     track(event: AnalyticsEvent): Promise<void>;
-    track(
+    async track(
         arg1: string | AnalyticsEvent,
         arg2?: Record<string, string | number | boolean>
     ): Promise<void> {
-        if (typeof arg1 === 'string') {
-            return this.trackEvent(arg1.toLowerCase(), {
-                ...this.user_properties,
-                ...(arg2 ?? {})
-            });
-        } else {
-            const { name, ...props } = arg1;
-            return this.trackEvent(name, { ...this.user_properties, ...props });
-        }
-    }
+        const eventName = typeof arg1 === 'string' ? arg1.toLowerCase() : arg1.name;
+        const eventProps =
+            typeof arg1 === 'string'
+                ? arg2 ?? {}
+                : (() => {
+                      const { name, ...rest } = arg1;
+                      return rest as Record<string, string | number | boolean>;
+                  })();
 
-    private async trackEvent(
-        eventName: string,
-        props?: Record<string, string | number | boolean>
-    ): Promise<void> {
-        try {
-            const { sessionId, ...identityProps } = await getUserIdentityProps(this.userIdentity);
+        const { sessionId, ...identityProps } = await getUserIdentityProps(this.userIdentity);
 
-            const response = await fetch(this.apiUrl, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'App-Key': this.appKey
-                },
-                credentials: 'omit',
-                body: JSON.stringify({
-                    timestamp: new Date().toISOString(),
-                    eventName,
-                    sessionId,
-                    systemProps: {
-                        locale: this.getBrowserLocale(),
-                        isDebug: this.isDebug,
-                        appVersion: this.appVersion,
-                        sdkVersion: 'custom@0.0.4',
-                        osName: this.getUserOS()
-                    },
-                    props: {
-                        ...props,
-                        ...identityProps
-                    }
-                })
-            });
-
-            if (response.status >= 300) {
-                const responseBody = await response.text();
-                console.warn(
-                    `Failed to send event "${eventName}": ${response.status} ${responseBody}`
-                );
-            }
-        } catch (e) {
-            console.warn(`Failed to send event "${eventName}"`);
-            console.warn(e);
-        }
-    }
-
-    private getBrowserLocale(): string | undefined {
-        if (typeof navigator === 'undefined') {
-            return undefined;
-        }
-
-        if (navigator.languages.length > 0) {
-            return navigator.languages[0];
-        } else {
-            return navigator.language;
-        }
-    }
-
-    private isDebug(): boolean {
-        return process.env.NODE_ENV === 'development';
-    }
-
-    private getUserOS() {
-        if (navigator.userAgent.includes('Win')) {
-            return 'Windows';
-        }
-        if (
-            navigator.userAgent.includes('iPhone') ||
-            navigator.userAgent.includes('iPad') ||
-            navigator.userAgent.includes('iPod')
-        ) {
-            return 'iOS';
-        }
-        if (navigator.userAgent.includes('Mac')) {
-            return navigator.maxTouchPoints > 1 ? 'iOS' : 'macOS';
-        }
-        if (navigator.userAgent.includes('Linux')) {
-            return 'Linux';
-        }
-        if (navigator.userAgent.includes('Android')) {
-            return 'Android';
-        }
-
-        return 'Unknown';
+        return trackEvent(eventName, {
+            ...this.user_properties,
+            ...eventProps,
+            ...identityProps,
+            app_session_id: sessionId
+        });
     }
 }
